+++ assets/js/wpdk.js
@@ -636,65 +636,87 @@ jQuery( function ( $ )
 {
   "use strict";
 
-  /**
-   * jQuery Cookie Plugin v1.3.1
+  /*!
+   * jQuery Cookie Plugin v1.4.1
    * https://github.com/carhartl/jquery-cookie
    *
-   * Copyright 2013 Klaus Hartl
+   * Copyright 2006, 2014 Klaus Hartl
    * Released under the MIT license
    */
-  (function ( factory ) { if ( typeof define === 'function' && define.amd ) {
-      // AMD. Register as anonymous module.
+  (function ( factory )
+  {
+    if ( typeof define === 'function' && define.amd ) {
+      // AMD
       define( ['jquery'], factory );
     }
+    else if ( typeof exports === 'object' ) {
+      // CommonJS
+      factory( require( 'jquery' ) );
+    }
     else {
-      // Browser globals.
+      // Browser globals
       factory( jQuery );
     }
-  }( function ( $ ) {
+  }( function ( $ )
+  {
 
     var pluses = /\+/g;
 
-    function raw( s )
+    function encode( s )
     {
-      return s;
+      return config.raw ? s : encodeURIComponent( s );
     }
 
-    function decoded( s )
+    function decode( s )
     {
-      return decodeURIComponent( s.replace( pluses, ' ' ) );
+      return config.raw ? s : decodeURIComponent( s );
     }
 
-    function converted( s )
+    function stringifyCookieValue( value )
+    {
+      return encode( config.json ? JSON.stringify( value ) : String( value ) );
+    }
+
+    function parseCookieValue( s )
     {
       if ( s.indexOf( '"' ) === 0 ) {
-        // This is a quoted cookie as according to RFC2068, unescape
+        // This is a quoted cookie as according to RFC2068, unescape...
         s = s.slice( 1, -1 ).replace( /\\"/g, '"' ).replace( /\\\\/g, '\\' );
       }
+
       try {
+        // Replace server-side written pluses with spaces.
+        // If we can't decode the cookie, ignore it, it's unusable.
+        // If we can't parse the cookie, ignore it, it's unusable.
+        s = decodeURIComponent( s.replace( pluses, ' ' ) );
         return config.json ? JSON.parse( s ) : s;
-      } catch (er) {
+      } catch (e) {
       }
+    }
+
+    function read( s, converter )
+    {
+      var value = config.raw ? s : parseCookieValue( s );
+      return $.isFunction( converter ) ? converter( value ) : value;
     }
 
     var config = $.cookie = function ( key, value, options )
     {
 
-      // write
-      if ( value !== undefined ) {
+      // Write
+
+      if ( arguments.length > 1 && !$.isFunction( value ) ) {
         options = $.extend( {}, config.defaults, options );
 
         if ( typeof options.expires === 'number' ) {
           var days = options.expires, t = options.expires = new Date();
-          t.setDate( t.getDate() + days );
+          t.setTime( +t + days * 864e+5 );
         }
 
-        value = config.json ? JSON.stringify( value ) : String( value );
-
         return (document.cookie = [
-          config.raw ? key : encodeURIComponent( key ),
+          encode( key ),
           '=',
-          config.raw ? value : encodeURIComponent( value ),
+          stringifyCookieValue( value ),
           options.expires ? '; expires=' + options.expires.toUTCString() : '', // use expires attribute, max-age is not supported by IE
           options.path ? '; path=' + options.path : '',
           options.domain ? '; domain=' + options.domain : '',
@@ -702,22 +724,29 @@ jQuery( function ( $ )
         ].join( '' ));
       }
 
-      // read
-      var decode = config.raw ? raw : decoded;
-      var cookies = document.cookie.split( '; ' );
+      // Read
+
       var result = key ? undefined : {};
+
+      // To prevent the for loop in the first place assign an empty array
+      // in case there are no cookies at all. Also prevents odd result when
+      // calling $.cookie().
+      var cookies = document.cookie ? document.cookie.split( '; ' ) : [];
+
       for ( var i = 0, l = cookies.length; i < l; i++ ) {
         var parts = cookies[i].split( '=' );
         var name = decode( parts.shift() );
-        var cookie = decode( parts.join( '=' ) );
+        var cookie = parts.join( '=' );
 
         if ( key && key === name ) {
-          result = converted( cookie );
+          // If second argument (value) is a function it's a converter...
+          result = read( cookie, value );
           break;
         }
 
-        if ( !key ) {
-          result[name] = converted( cookie );
+        // Prevent storing a cookie that we couldn't decode.
+        if ( !key && (cookie = read( cookie )) !== undefined ) {
+          result[name] = cookie;
         }
       }
 
@@ -728,12 +757,13 @@ jQuery( function ( $ )
 
     $.removeCookie = function ( key, options )
     {
-      if ( $.cookie( key ) !== undefined ) {
-        // Must not alter options, thus extending a fresh object...
-        $.cookie( key, '', $.extend( {}, options, { expires : -1 } ) );
-        return true;
+      if ( $.cookie( key ) === undefined ) {
+        return false;
       }
-      return false;
+
+      // Must not alter options, thus extending a fresh object...
+      $.cookie( key, '', $.extend( {}, options, {expires : -1} ) );
+      return !$.cookie( key );
     };
 
   } ));