+++ CHANGELOG.md
@@ -20,6 +20,17 @@ For more information on SemVer, please visit http://semver.org/.
 
 ---
 
+## Version 1.5.12
+### 2014-08-??
+
+#### Improvements
+
+* Updated jQuery Cookie Plugin to v1.4.1 (https://github.com/carhartl/jquery-cookie)
+
+#### Bugs
+
+* Fixed jQuery UI css conflict
+
 ## Version 1.5.11
 ### 2014-08-20